move nested windmill state classes to top level files and organize directory paths
